added remove from cart and purchase on frontend

+++ MusicStoreApi/Handlers/Customers/Commands/RemoveCartItem.cs
@@ -18,10 +18,12 @@ namespace MusicStoreApi.Handlers.Customers.Commands
         public class RequestHandler : IRequestHandler<Command, bool>
         {
             private readonly IRepository<CartItem> _cartItemRepository;
+            private readonly IRepository<Cart> _cartRepository;
 
-            public RequestHandler(IRepository<CartItem> cartItemRepository)
+            public RequestHandler(IRepository<CartItem> cartItemRepository, IRepository<Cart> cartRepository)
             {
                 _cartItemRepository = cartItemRepository ?? throw new ArgumentNullException(nameof(cartItemRepository));
+                _cartRepository = cartRepository ?? throw new ArgumentNullException(nameof(_cartRepository));
             }
             public Task<bool> Handle(Command request, CancellationToken cancellationToken)
             {
@@ -36,8 +38,18 @@ namespace MusicStoreApi.Handlers.Customers.Commands
                     throw new EntityDoesntExistException();
                 }
 
+                var cart = _cartRepository.GetById(cartItem.Cart.Id);
+
+                if (cart == null)
+                {
+                    throw new EntityDoesntExistException();
+                }
+
+                cart.CartValue -= cartItem.Quantity * cartItem.Product.Price;
                 _cartItemRepository.Delete(cartItem);
+                
                 _cartItemRepository.SaveChanges();
+                _cartRepository.SaveChanges();
 
                 return Task.FromResult(true);
             }
